Replaced MathOp with OperatorOp
(cherry picked from commit 63e4877)
