+++ _operations.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from collections import OrderedDict
 from _internalobjects import StackObj
 # TODO: Documentazione
 
@@ -23,6 +24,42 @@ class NestedOperation(ABC):
 
     def OutputRegKey(self):
         return self._OREGKEY
+
+
+class OperatorOperation(Operation, NestedOperation):
+    def __init__(self, name1, name2, nreg, reservebits, exitreg):
+        super().__init__()
+        self._name1 = name1
+        self._name2 = name2
+        self._nreg = nreg  # Numero di registri
+        self._exitreg = int(exitreg)  # Registro d'uscita (numerazione interna 0,1,2,...)
+        self._reservebits = reservebits  # Bit di riserva post operazione
+        self._choosedreg = OrderedDict()
+
+    def PreCompile(self, env):  # TODO: Somma Nestata
+        if self._name1 not in env.StackObject:
+            raise Exception("Variabile non definita")
+        if self._name2 not in env.StackObject:
+            raise Exception("Variabile non definita")
+        # Registry
+        choosedreg = self._choosedreg
+        nreg = self._nreg
+        for regKey, regObj in env.RegistryColl.items():
+            if not regObj.ReserveBit and len(choosedreg) < nreg:
+                choosedreg[regKey] = regObj
+            elif len(choosedreg) == nreg:
+                break
+        while len(choosedreg) < nreg:
+            regkey, regobj = env.RequestRegistry()
+            choosedreg[regkey] = regobj
+        vals = list(choosedreg.values())
+        for i in range(len(vals)):
+            vals[i].ReserveBit = self._reservebits[i]
+        self._OREGKEY = tuple(choosedreg.keys())[self._exitreg]
+
+    @abstractmethod
+    def GetCode(self, env, p):
+        pass
 
 
 # TODO: creare un sistema per la memorizzazione tipizzata
